refactor: normalize page filenames to further match featurePage pattern standards
