+++ src/client/pages/editorPage/editorPage.tsx
