+++ src/client/pages/researchPage/researchPage.tsx
